mcp-servers/plantation-mcp/pyproject.toml
[tool.poetry]
name = "plantation-mcp"
version = "0.1.0"
description = "Plantation MCP Server - Provides farmer and collection point data to AI agents"
authors = ["Farmer Power Team"]
readme = "README.md"
packages = [{include = "plantation_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.60.0"
grpcio-health-checking = "^1.60.0"
grpcio-reflection = "^1.60.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
structlog = "^24.1.0"
tenacity = "^8.2.0"
jsonschema = "^4.20.0"
# OpenTelemetry
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-otlp = "^1.29.0"
opentelemetry-instrumentation-grpc = ">=0.50b0"
# Local dependencies
fp-common = { path = "../../libs/fp-common" }
fp-proto = { path = "../../libs/fp-proto" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
fp-testing = { path = "../../libs/fp-testing" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
